+++ lib/workato/connector/sdk/errors.rb
@@ -6,11 +6,50 @@ module Workato
     module Sdk
       InvalidDefinitionError = Class.new(StandardError)
 
-      InvalidSchemaError = Class.new(StandardError)
+      class UnexpectedMethodDefinitionError < InvalidDefinitionError
+        attr_reader :name
+        attr_reader :definition
+
+        def initialize(name, definition)
+          super("Expected lambda for method '#{name}' definition, got: #{definition.class.name}")
+          @name = name
+          @definition = definition
+        end
+      end
+
+      class UndefinedMethodError < InvalidDefinitionError
+        attr_reader :name
+
+        def initialize(name)
+          super("Method '#{name}' does not exists")
+          @name = name
+        end
+      end
+
+      InvalidSchemaError = Class.new(InvalidDefinitionError)
 
       CustomRequestError = Class.new(StandardError)
 
       RuntimeError = Class.new(StandardError)
+
+      class UnresolvedObjectDefinitionError < StandardError
+        attr_reader :name
+
+        def initialize(name)
+          super("Cannot find object definition for '#{name}'")
+          @name = name
+        end
+      end
+
+      class CircleReferenceObjectDefinitionError < StandardError
+        attr_reader :name
+
+        def initialize(name, backtrace = [])
+          super("Infinite recursion occurred in object definition for '#{name}'")
+          set_backtrace(backtrace)
+          @name = name
+        end
+      end
 
       class RequestError < StandardError
         attr_reader :method
@@ -41,6 +80,24 @@ module Workato
           super(message)
         end
       end
+
+      RequestTLSCertificateFormatError = Class.new(StandardError)
+
+      RequestPayloadFormatError = Class.new(StandardError)
+
+      JSONRequestFormatError = Class.new(RequestPayloadFormatError)
+
+      JSONResponseFormatError = Class.new(RequestPayloadFormatError)
+
+      XMLRequestFormatError = Class.new(RequestPayloadFormatError)
+
+      XMLResponseFormatError = Class.new(RequestPayloadFormatError)
+
+      WWWFormURLEncodedRequestFormatError = Class.new(RequestPayloadFormatError)
+
+      MultipartFormRequestFormatError = Class.new(RequestPayloadFormatError)
+
+      RAWResponseFormatError = Class.new(RequestPayloadFormatError)
     end
   end
 end